Move file/directory types to the types directory.
