+++ bids-validator/src/compat/fulltest.ts
@@ -1,6 +1,6 @@
 /** Adapter to run Node.js bids-validator fullTest with minimal changes from Deno */
 import { ValidatorOptions } from '../setup/options.ts'
-import { FileTree } from '../files/filetree.ts'
+import { FileTree } from '../types/filetree.ts'
 import { walkFileTree } from '../schema/walk.ts'
 import { Issue } from '../types/issues.ts'
 import validate from '../../dist/esm/index.js'
