+++ bids-validator/src/validators/bids.ts
@@ -1,5 +1,4 @@
-import { Issue } from '../types/issues.ts'
-import { FileTree } from '../files/filetree.ts'
+import { FileTree } from '../types/filetree.ts'
 import { walkFileTree } from '../schema/walk.ts'
 import { loadSchema } from '../setup/loadSchema.ts'
 import { applyRules } from '../schema/applyRules.ts'
